- use correct size type, add missing ftruncate() return value check (Ralf)

+++ apt-pkg/contrib/mmap.h
@@ -62,7 +62,7 @@ class MMap
    
    // File manipulators
    bool Sync();
-   bool Sync(unsigned long Start,unsigned long Stop);
+   bool Sync(size_t Start,size_t Stop);
    
    MMap(FileFd &F,unsigned long Flags);
    MMap(unsigned long Flags);
